+++ tests/js/client/load-balancing/load-balancing-queries-noauth-cluster.js
@@ -32,26 +32,7 @@ const db = require("internal").db;
 const request = require("@arangodb/request");
 const url = require('url');
 const _ = require("lodash");
-
-function getCoordinators() {
-  const isCoordinator = (d) => (_.toLower(d.instanceRole) === 'coordinator');
-  const toEndpoint = (d) => (d.endpoint);
-  const endpointToURL = (endpoint) => {
-    if (endpoint.substr(0, 6) === 'ssl://') {
-      return 'https://' + endpoint.substr(6);
-    }
-    var pos = endpoint.indexOf('://');
-    if (pos === -1) {
-      return 'http://' + endpoint;
-    }
-    return 'http' + endpoint.substr(pos);
-  };
-
-  const instanceInfo = JSON.parse(require('internal').env.INSTANCEINFO);
-  return instanceInfo.arangods.filter(isCoordinator)
-                              .map(toEndpoint)
-                              .map(endpointToURL);
-}
+const getCoordinators = require('@arangodb/test-helper').getCoordinators;
 
 const servers = getCoordinators();
